+++ files/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [typescript v2.0.0] files
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@wingspan/files/v/2.0.0 - files
+
+## 2024-02-29 16:18:58
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f51b81948601a18934c5
+- Speakeasy CLI 1.197.2 (2.275.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.1] files
+### Releases
+- [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/files/v/2.0.1 - files